optionally enable MPI in hyperion SED call
The mpi flag on `m.run` means that the MPI version of hyperion is called. When running this interactively on some systems (on a single node) this fails, even if `mpi_processes` is set to 1, as it contravenes some system policies. This changes avoids this by checking if a multi-node job has been specified